+++ main.rb
@@ -6,18 +6,9 @@ require_relative 'game_class'
 require_relative 'genre'
 require_relative 'label'
 require_relative 'music_album'
+require_relative 'app'
 
-def display_options
-  puts 'Options:'
-  puts '1. Add a new book'
-  puts '2. Add a new game'
-  puts '3. Add a new music album'
-  puts '4. List all items'
-  puts '5. List music_albums'
-  puts '6. List all books'
-  puts '7. List all games'
-  puts '8. Quit'
-end
+app = App.new
 
 # Store your items
 items = []
@@ -137,7 +128,7 @@ def display_game(index, game)
 end
 
 loop do
-  display_options
+  app.display_options
   print 'Select an option: '
   choice = gets.chomp.to_i
